Basic Combat Implementation Finished
I have the enemy damage the player and the player damage the enemy. This works mostly well the only things I want to change is adding more feedback when the player gets hit as well as when the enemy gets hit. This will make the game feel more fluid and fun. Both entities don't have any functional health or indication that they have been hit (outside of the console log)