Sign Creates Json File
My sign creates a JSON file storing all the data that need to be read in the file. The file is saved as two seporate files one in the unity drive and one on my google drive. I need to figure out how to make both the same as the one saved in my google drive saves a new one for every sign data created where as the one in my google drive add the data to the file.